+++ map/src/main/java/com/adsamcik/tracker/map/heatmap/creators/WifiCountHeatmapTileCreator.kt
@@ -26,7 +26,7 @@ internal class WifiCountHeatmapTileCreator(context: Context, val layerData: MapL
 				{ current, _, stampValue, weight ->
 					current + stampValue * weight
 				}) { current, stampValue, weight ->
-			((current.toFloat() + stampValue * weight) / 2f).toInt().toUByte()
+			((current.toFloat() + stampValue * weight) / 2f).toInt()
 		}
 	}
